requirements.txt
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
dotenv-python==0.0.1
Flask==2.2.2
Flask-Login==0.6.2
flask-ngrok==0.0.25
Flask-WTF==1.1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.3.0
pycodestyle==2.10.0
python-dotenv==0.21.0
qrcode==7.3.1
requests==2.28.1
tomli==2.0.1
urllib3==1.26.12
Werkzeug==2.2.2
WTForms==3.0.1
